Reworked to not use the deprecated IndexModifier
This moves to using IndexWriter instead of IndexModifier for all update
and delete operations.  As a result of the change the docs replaced and
docs deleted counts are no longer possible and have been removed.  This
seems like a fine change as the focus should be on terms and fields.

While working through all these changes one of the largest problems
seemed to be the constant use of switching between compound files and
then optimizing with every single operation.  To alleviate this the
defalut clucene settings are now used until a call is made to optimize()
on the lucene object.  This call will switch the entire index to
compound files and run a full optimization pass.